restructured loading animation
