+++ src/pages/MyAccountPage/MyAccount.jsx
@@ -2,7 +2,7 @@ import { useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link, Navigate } from "react-router-dom";
 import animation from "../../animations/popupbox";
-import LoadingScreen from "../../components/loadingScreen";
+import LoadingScreen from "../../components/Loading/loadingScreen";
 import Modal from "../../components/modal";
 import { apiURL } from "../../store/actions";
 import { todolistSlice } from "../../store/todolistSlice";
